Update Migration File
